+++ RockPaperScissors/RockPaperScissors/Game/RPSGame.swift
@@ -1,5 +1,5 @@
 //
-//  RPSGameIteration.swift
+//  RPSGame.swift
 //  RockPaperScissors
 //
 //  Created by Effie on 11/28/23.
@@ -7,36 +7,7 @@
 
 import Foundation
 
-enum RPSResult {
-    case win(RPSPlayable)
-    case draw
-}
-
-struct RPSPart {
-    private let leftGesture: RPSGesture
-    
-    private let rightGesture: RPSGesture
-    
-    init(between leftGesture: RPSGesture, and rightGesture: RPSGesture) {
-        self.leftGesture = leftGesture
-        self.rightGesture = rightGesture
-    }
-    
-    func getResult() -> RPSResult {
-        let leftHand = leftGesture.hand
-        let rightHand = rightGesture.hand
-        
-        if leftHand == rightHand {
-            return .draw
-        } else {
-            let winningGesture = leftHand.wins(rightHand) ? leftGesture : rightGesture
-            let winner = winningGesture.owner
-            return .win(winner)
-        }
-    }
-}
-
-struct RPSIteration {
+struct RPSGame {
     private let leftPlayer: RPSPlayable
     
     private let rightPlayer: RPSPlayable
