+++ RockPaperScissors/RockPaperScissors/IO/Console.swift
@@ -1,31 +1,11 @@
 //
-//  IO.swift
+//  Console.swift
 //  RockPaperScissors
 //
-//  Created by Effie on 11/28/23.
+//  Created by Effie on 11/29/23.
 //
 
 import Foundation
-
-protocol PromptDisplayable {
-    func displayPrompt(_ prompt: String)
-}
-
-protocol InputGettable {
-    func getInput() throws -> String
-}
-
-protocol OuputDisplayble {
-    func displayOutput(_ output: String)
-}
-
-protocol RPSErrorDisplayble {
-    func displayRPSError(_ error: RPSError)
-}
-
-typealias Displayable = PromptDisplayable & OuputDisplayble & RPSErrorDisplayble
-
-typealias IO = InputGettable & Displayable
 
 final class Console: InputGettable {
     func getInput() throws -> String {
