+++ RockPaperScissors/RockPaperScissors/Result/MJBResult.swift
@@ -1,0 +1,13 @@
+//
+//  MJBResult.swift
+//  RockPaperScissors
+//
+//  Created by Effie on 11/29/23.
+//
+
+import Foundation
+
+enum MJBResult {
+    case win
+    case regame(nextTurn: MJBPlayable)
+}
