+++ RockPaperScissors/RockPaperScissors/Result/RPSResult.swift
@@ -1,0 +1,13 @@
+//
+//  RPSResult.swift
+//  RockPaperScissors
+//
+//  Created by Effie on 11/29/23.
+//
+
+import Foundation
+
+enum RPSResult {
+    case win(RPSPlayable)
+    case draw
+}
